+++ Button.h
@@ -5,11 +5,11 @@ class Button {
     int pin;
   	int state;
   	int lastReading;
-  	int lastDebounceTime = 0;
-    int debounceDelay = 50;
-    int lastStart = 0;
-    int lastEnd = 0;
-    int duration = 0;
+  	unsigned long lastDebounceTime = 0;
+    unsigned long debounceDelay = 50;
+    unsigned long lastStart = 0;
+    unsigned long lastEnd = 0;
+    unsigned long duration = 0;
   public:
     Button(int buttonPin);
   	void init();
@@ -19,11 +19,11 @@ class Button {
     void update();
     int getState();
     int getLastReading();
-    void setLastStart(int start);
-    int getLastStart();
-    void setLastEnd(int endTime);
-    int getLastEnd();
-    void setDuration(int dur);
-    int getDuration();
+    void setLastStart(unsigned long start);
+    unsigned long getLastStart();
+    void setLastEnd(unsigned long endTime);
+    unsigned long getLastEnd();
+    void setDuration(unsigned long dur);
+    unsigned long getDuration();
     bool isPressed();
 };
